Change MemoryAccessor#deallocateMemory to use a general pointer, which lets it be used properly (subclasses now have to override MemoryAccessor#deallocateOurMemory, and provide Class objects for their error messages and pointer types)

+++ src/main/java/org/abimon/colonelAccess/osx/structs/VMRegionSubmapInfo64.java
@@ -24,7 +24,7 @@ public class VMRegionSubmapInfo64 extends Structure {
     public boolean is_submap;
     public int behavior;
     public long object_id;
-    public int user_wired_count;
+    public short user_wired_count;
 
     private static final List<String> FIELD_NAMES = Arrays.asList(
             "protection",
